// src/config/enums.ts
export enum SOCKET_EVENTS {
    CONNECTION = "connection",
    DISCONNECT = "disconnect",
    ERROR = "error",
    IMAGE_UPLOAD_PROGRESS = "image_upload_progress",
    VIDEO_UPLOAD_PROGRESS = "video_upload_progress",
    PDF_EXTRACT_PROGRESS = "pdf_extract_progress",
    PDF_UPLOAD_PROGRESS = "pdf_upload_progress",
}